styling, sorting, other fixes

+++ client/src/components/EventCard.js
@@ -39,9 +39,22 @@ export class EventCard extends Component {
 
   }
 
-  onDelete(event_id) {
+  onDelete = (event_id) => {
     console.log(event_id)
-    this.props.deleteEvent(event_id, this.props.token)
+    const room = this.props.events.find(event => {
+      return event._id === event_id
+    });
+    if (!room.password) {
+      this.props.deleteEvent(event_id, this.props.token)
+    }
+    let password = prompt('Please enter the password here:');
+    if (password === null || password === '') {
+      return alert('must enter a value')
+    } else if (password === room.password) {
+      this.props.deleteEvent(event_id, this.props.token)
+    } else {
+      return alert('Incorrect Password!')
+    }
   }
 
   render() {
